Refactor components to use explicit internal imports
