+++ src/lib/modules/charts/components/extra-params/threshold.component.ts
@@ -1,12 +1,9 @@
 import { AfterViewInit, Component, EventEmitter, Input, Output, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
-import {
-  Indicator,
-  PrecipitationUnits,
-  ThresholdIndicatorQueryParams,
-  TemperatureUnits
-} from '../../../../public_api';
+import { Indicator } from '../../../api/models/indicator.model';
+import { PrecipitationUnits, TemperatureUnits } from '../../../shared/extra-params.constants';
+import { ThresholdIndicatorQueryParams } from '../../../api/models/threshold-indicator-query-params.model';
 
 /*
  * Threshold params component
